--- .env_example
@@ -1,2 +1,0 @@
-WEBHOOK=
-GITHUB_ACCESS_TOKEN=
