+++ README.md
@@ -1,13 +1,16 @@
+## 🖥️ Hosting
+[![Deploy with Vercel](https://vercel.com/button)](https://vercel.com/new/clone?repository-url=https%3A%2F%2Fgithub.com%2FJanjyTapYT%2Fportfolio&env=webhook,GITHUB_ACCESS_TOKEN&envDescription=You%20need%20the%20webhook%20variable%20to%20let%20the%20contact%20page%20work.%20you%20can%20create%20webhook%20url%20in%20a%20dscord%20server%20%2F%20channel%20and%20paste%20it%20in%20the%20variable%20value&project-name=portfolio&repo-name=portfolio&demo-title=JanjyTapYT's%20Website&demo-description=A%20cool%20designed%20developer%20website%20built%20with%20Nextjs%20and%20react.&demo-url=https%3A%2F%2Fjanjytapyt.me&demo-image=https%3A%2F%2Fcdn.discordapp.com%2Fattachments%2F971049189377179718%2F1034191567763816448%2Funknown.png)
+
 ## Getting Started
 
 ## 🛠 Development Server
 
 ```bash
-npm install --legacy-peer-deps && npm start dev
+npm install && npm start dev
 ```
 ## 🛠 Production Server
 ```bash
-npm install --legacy-peer-deps && npm run build && npm run start
+npm install && npm run build && npm run start
 ```
 
 ### Terms of Use
@@ -16,7 +19,7 @@ By using this website, you agree to the following terms of use:
 ```md
 - You are not allowed to sell this website or any part of it.
 - You are not allowed to claim this website as your own.
-- You are not allowed to change the credits ("Developed with ❤️ by BinaryBlazer") in the footer of the website.
+- You are not allowed to change the credits ("Developed with ❤️ by JanjyTapYT") in the footer of the website.
 - You are allowed to modify the code of this website as long as you don't sell it or claim it as your own.
 - You have the permission to share in writing, but you have to mention me, in the article.
 - You can't speak in a "i did it" way.
@@ -32,7 +35,7 @@ To learn more about Next.js, take a look at the following resources:
 You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
 
 ## If you want to display your Discord Presence on the site:
-#### you have to join the [lanyard discord server](https://discord.gg/G6pY5YjjTQ) for it to work. and make sure to enter your discord id in the ``./site.config.js`` at the line ``25``. 
+#### you have to join the [lanyard discord server](https://discord.gg/G6pY5YjjTQ) for it to work. and make sure to enter your discord id in the ``./site.config.js`` at the line ``18``. 
 
 
 ## ⭐ Star
@@ -42,5 +45,5 @@ You can check out [the Next.js GitHub repository](https://github.com/vercel/next
 ## 💫 Credits
 
 ```bash
-Developed with ❤️ by BinaryBlazer
+Developed with ❤️ by JanjyTapYT
 ```
