+++ package.json
@@ -1,50 +1,29 @@
 {
   "name": "portfolio",
-  "version": "2.1.0",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "start": "next start",
-    "test": "echo \"Success!\""
+    "start": "next start"
   },
   "dependencies": {
     "@headlessui/react": "^1.7.3",
-    "@next/mdx": "^13.2.4",
-    "@nextui-org/react": "^1.0.0-beta.12",
-    "@octokit/rest": "^19.0.7",
-    "@tailwindcss/postcss7-compat": "^2.2.17",
     "@tippyjs/react": "^4.2.6",
     "axios": "^1.1.3",
     "framer-motion": "^5.2.1",
-    "gray-matter": "^4.0.3",
-    "memfs": "^3.4.13",
-    "memory-fs": "^0.5.0",
-    "next": "^13.3.0",
-    "next-mdx-remote": "^4.4.1",
+    "next": "^12.3.1",
     "next-pwa": "^5.6.0",
-    "nprogress": "^0.2.0",
-    "path": "^0.12.7",
-    "postcss-import": "^12.0.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-markdown": "^8.0.6",
-    "react-parallax-tilt": "^1.7.118",
     "react-particles": "^2.5.3",
-    "react-text-loop-next": "^0.0.3",
-    "remark-mdx": "^2.3.0",
-    "remark-stringify": "^10.0.2",
-    "@sdevs/screen-sizr": "^2.0.0",
     "swr": "^1.3.0",
-    "to-vfile": "^7.2.4",
     "tsparticles": "^2.5.3"
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.7",
-    "autoprefixer": "^9.8.6",
+    "autoprefixer": "^10.4.11",
     "eslint": "8.23.1",
     "eslint-config-next": "12.3.0",
-    "postcss": "^7.0.35",
+    "postcss": "^8.4.16",
     "tailwindcss": "^3.1.8"
   }
 }
