--- src/data/blog/posts/welcome/content.md
@@ -1,5 +1,0 @@
-Hi there! Its 19:16 GMT + 1 right now. I'm currently working on the blog system. I'm unsure if I should use a database or just use markdown files. I'm leaning towards markdown files because it's easier to edit and I don't have to worry about a database.
-
-But anyways, you are currently reading the first post on my blog. I'm not sure what to write about, so I'll just write about what I'm doing right now. I'm currently working on the blog system and when im done i will play War Thunder so lets see what i will do after playing War Thunder.
-
-Anyways thank you for reading and have a nice day!
