--- src/styles/tippy.css
@@ -1,99 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Karla:200,300,regular,500,600,700,800,200italic,300italic,italic,500italic,600italic,700italic,800italic");
-
-.tippy-box{
-    position: relative;
-    font-family: 'Karla', sans-serif;
-    background-color:#333;
-    color:#fff;
-    border-radius:4px;
-    font-size:14px;
-    line-height:1.4;
-    outline:0;
-    transition-property:transform,visibility,opacity
-}
-.tippy-box[data-placement^=top]>.tippy-arrow{
-    bottom:0
-}
-.tippy-box[data-placement^=top]>.tippy-arrow:before{
-    bottom:-7px;
-    left:0;
-    border-width:8px 8px 0;
-    border-top-color:initial;
-    transform-origin:center top
-}
-.tippy-box[data-placement^=bottom]>.tippy-arrow{
-    top:0
-}
-.tippy-box[data-placement^=bottom]>.tippy-arrow:before{
-    top:-7px;
-    left:0;
-    border-width:0 8px 8px;
-    border-bottom-color:initial;
-    transform-origin:center bottom
-}
-.tippy-box[data-placement^=left]>.tippy-arrow{
-    right:0
-}
-.tippy-box[data-placement^=left]>.tippy-arrow:before{
-    border-width:8px 0 8px 8px;
-    border-left-color:initial;
-    right:-7px;
-    transform-origin:center left
-}
-.tippy-box[data-placement^=right]>.tippy-arrow{
-    left:0
-}
-.tippy-box[data-placement^=right]>.tippy-arrow:before{
-    left:-7px;
-    border-width:8px 8px 8px 0;
-    border-right-color:initial;
-    transform-origin:center right
-}
-.tippy-box[data-inertia][data-state=visible]{
-    transition-timing-function:cubic-bezier(.54,1.5,.38,1.11)
-}
-.tippy-arrow{
-    width:16px;
-    height:16px;
-    color:#333
-}
-.tippy-arrow:before{
-    content:"";
-    position:absolute;
-    border-color:transparent;
-    border-style:solid
-}
-.tippy-content{
-    position:relative;
-    padding:5px 9px;
-    z-index:1
-}
-
-.tippy-box {
-  background-color: #000;
-  color: #fff;
-  transition: all 0.1s;
-  box-shadow: 0px 0px 25px #000;
-  outline: #0055ff solid 2px;
-  font-size: 16px;
-  font-weight: 600;
-}
-.tippy-arrow {
-  color: #000;
-}
-.tippy-box[data-placement^=top]{
-    transform-origin:bottom
-}
-.tippy-box[data-placement^=bottom]{
-    transform-origin:top
-}
-.tippy-box[data-placement^=left]{
-    transform-origin:right
-}
-.tippy-box[data-placement^=right]{
-    transform-origin:left
-}
-.tippy-box[data-state=hidden]{
-    transform:scale(.5);
-    opacity:0
-}
